FafCarsApi/Scripts/py/requirements.txt
et-xmlfile==1.1.0
numpy==2.0.0
openpyxl==3.1.5
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
